a
a

+++ admin/a_jobs/ftp.php
@@ -5,8 +5,11 @@
   		parent::__construct();
   	}
   }
-  $destino="file.json";
+  $destino="file.json";     //////////////////////ARCHIVO JSON
   echo "<br>".date("Y-m-d H:i:s");
+
+  ///////////////////////////////////////CLASE PARA EL FTP
+  echo "<br>CLASE FTP";
   class ftp{
       public $conn;
 
@@ -42,36 +45,37 @@
       }
   }
 
+  //////////////////////////////////////////////DESCARGA EL ARCHIVO JSON
   $ftp = new ftp('216.70.82.104');
   $ftp->ftp_login('PAC0736','Eku3q1dxIAbXe2u39Jtv');
   if (file_exists ($destino)){
-    unlink($destino);
+    unlink($destino);          //////////////////SE BORRA EL ANTERIOR JSON EN CASO DE EXISTIR
   }
-
   $ftp->descargar("catalogo_xml/productos.json",$destino,FTP_BINARY);
   $ftp->desconectar();
-  $fmodif=date("Y-m-d H:i:s");
+  echo "<br>DESCARGA DEL ARCHIVO";
+
+  $fmodif=date("Y-m-d H:i:s");    ///////////////   HORA DE MOFIDICACION
+  $update_hora=date("H");         ///////////////   VARIABLE PARA ACTUALIZAR
+  /////////////////////////////////SI SE DESCARGA EL JSON HAY PROCESO
+  echo "<br>HORA PROCESO ".$fmodif;
+
   if (file_exists ($destino)){
     ///////////////////////////////////////    PROCESO  /////////////////////////////////////////////////////////////
     $db = new Productos();
-/*
-    $sql="TRUNCATE TABLE producto_exist";
-    $sth0 = $db->dbh->prepare($sql);
-    echo "Truncate:".$sth0->execute();
 
-    $sql="ALTER TABLE producto_exist AUTO_INCREMENT = 1";
-    $sth0 = $db->dbh->prepare($sql);
-    $sth0->execute();
-*/
-
+    //////////////////////////////////////   JSON SE CONVIERTE EN ARREGLO
     $data = file_get_contents($destino);
     $products = json_decode($data, true);
     $x="";
+    echo "<br>A ARREGLO";
 
     $i=0;
+    ////////////////////////////////////     SE RECORRE CADA PRODUCTO DEL JSON
     foreach ($products as $product) {
         try{
-          $sql="select * from productos where idProducto='".$product['idProducto']."' limit 1";
+          ////////////////////////BUSCAMOS EL PRODUCTO DEL JSON SI EXISTE SOLO SE ACTUALIZA, SI NO EXISE SE INGRESA
+          $sql="select idProducto from productos where idProducto='".$product['idProducto']."' limit 1";
           $stmt= $db->dbh->query($sql);
           if($stmt->rowCount()==0){
             $sql="insert into productos (idProducto, clave, numParte, nombre, modelo, idMarca, marca, idCategoria, categoria, idSubCategoria, subcategoria, descripcion_corta, precio, moneda, tipoCambio, imagen, upc, activo, modificado) values (:idProducto, :clave, :numParte, :nombre, :modelo, :idMarca, :marca, :idCategoria, :categoria, :idSubCategoria, :subcategoria, :descripcion_corta, :precio, :moneda, :tipoCambio, :imagen, :upc, :activo, :modificado)";
@@ -95,6 +99,8 @@
             $sql="update productos set precio=:precio, moneda=:moneda, tipoCambio=:tipoCambio, imagen=:imagen, activo=:activo, modificado=:modificado where idProducto='".$product['idProducto']."'";
             $sth = $db->dbh->prepare($sql);
           }
+
+          //////////////////////////// SOLO SE ACTUALIZAN UNOS CAMPOS
           $sth->bindValue(':precio', $product['precio']);
           $sth->bindValue(':moneda', $product['moneda']);
           $sth->bindValue(':tipoCambio', $product['tipoCambio']);
@@ -103,53 +109,80 @@
           $sth->bindValue(':modificado', $fmodif);
           $sth->execute();
 
+          //////////////////////////// SE BORRAN LAS EXISTENCIAS DEL PRODUCTO PARA ACTUALIZARLAS
           $sql="delete from producto_exist where idProducto='".$product['idProducto']."'";
           $sth3 = $db->dbh->prepare($sql);
           $sth3->execute();
 
-          while (current($product['existencia'])) {
-            $name=key($product['existencia']);
-            $valor=$product['existencia'][$name];
+          ///////////////////////////  SE ACTUALIZAN LAS EXISTENCIAS CADA HORA
+          if (is_array($product['existencia'])){
+            while (current($product['existencia'])) {
+              $name=key($product['existencia']);
+              $valor=$product['existencia'][$name];
 
-            $sql="insert into producto_exist (idProducto, almacen, existencia) values (:idProducto, :almacen, :existencia)";
-            $sth2 = $db->dbh->prepare($sql);
-            $sth2->bindValue(':idProducto', $product['idProducto']);
-            $sth2->bindValue(':almacen', $name);
-            $sth2->bindValue(':existencia', $valor);
-            $sth2->execute();
-            next($product['existencia']);
+              $sql="insert into producto_exist (idProducto, almacen, existencia) values (:idProducto, :almacen, :existencia)";
+              $sth2 = $db->dbh->prepare($sql);
+              $sth2->bindValue(':idProducto', $product['idProducto']);
+              $sth2->bindValue(':almacen', $name);
+              $sth2->bindValue(':existencia', $valor);
+              $sth2->execute();
+              next($product['existencia']);
+            }
           }
 
-          $sql="delete from producto_espe where idProducto='".$product['idProducto']."'";
-          $sth3 = $db->dbh->prepare($sql);
-          $sth3->execute();
-
-          foreach($product['especificaciones'] as $key){
-            $sql="insert into producto_espe (idProducto, tipo, valor) values (:idProducto, :tipo, :valor)";
-            $sth2 = $db->dbh->prepare($sql);
-            $sth2->bindValue(':idProducto', $product['idProducto']);
-            $sth2->bindValue(':tipo', $key['tipo']);
-            $sth2->bindValue(':valor', $key['valor']);
-            $sth2->execute();
+          ////////////////////////////  SE ACTUALIZAN LAS ESPECIFICACIONES A LAS 8 DE LA MAÑANA Y LAS 8 DE LA NOCHE
+          if ($update_hora==8 OR $update_hora=20){
+            $sql="delete from producto_espe where idProducto='".$product['idProducto']."'";
+            $sth3 = $db->dbh->prepare($sql);
+            $sth3->execute();
+            if (is_array($product['especificaciones'])){
+              foreach($product['especificaciones'] as $key){
+                $sql="insert into producto_espe (idProducto, tipo, valor) values (:idProducto, :tipo, :valor)";
+                $sth2 = $db->dbh->prepare($sql);
+                $sth2->bindValue(':idProducto', $product['idProducto']);
+                $sth2->bindValue(':tipo', $key['tipo']);
+                $sth2->bindValue(':valor', $key['valor']);
+                $sth2->execute();
+              }
+            }
           }
-
         }
         catch(PDOException $e){
           return "Database access FAILED! ".$e->getMessage();
         }
         $i++;
-/*
+        /*
         if ($i==2){
           break;
         }
-*/
+        */
     }
 
-    /////////////////////////////////////////////////////////////////////////////////////////////////////
+    /////////////////////////////////////// BUSCAMOS PRODUCTOS QUE NO SE ACTUALIZARON PARA ELIMINAR EXISTENCIAS
+    $sql="select * from productos where modificado!='$fmodif'";
+    $stmt= $db->dbh->query($sql);
+
+    ///////////// se borran existencias
+    $sql="delete from producto_exist where idProducto=:idproducto";
+    $sth3 = $db->dbh->prepare($sql);
+
+    ///////////// se desactivan los productos
+    $sql="update productos set activo=0 where idProducto=:idproducto";
+    $sth4 = $db->dbh->prepare($sql);
+
+    foreach($stmt as $key){
+      $sth3->bindValue(':idproducto', $key['idProducto']);
+      $sth3->execute();
+
+      $sth4->bindValue(':idproducto', $key['idProducto']);
+      $sth4->execute();
+    }
+
+    $date=date("YmdHis");
+    $file="file_".$date.".json";
+    rename($destino, "../historial/$file");
+    echo "<br>".date("Y-m-d H:i:s");
   }
 
-  $date=date("YmdHis");
-  $file="file_".$date.".json";
-  rename($destino, "../historial/$file");
-  echo "<br>".date("Y-m-d H:i:s");
+
 ?>
